Форма экрана авторизации вынесена
